# Greatly speed up homebrew module when multiple packages are passed in the `name` key 