--- a/changelogs/fragments/9181-improve-homebrew-module-performance.yml
+++ b/changelogs/fragments/9181-improve-homebrew-module-performance.yml
@@ -1,2 +1,2 @@
 minor_changes:
-  - homebrew - greatly speed up homebrew module when multiple packages are passed in the `name` key (https://github.com/ansible-collections/community.general/pull/9181).
+  - homebrew - greatly speed up homebrew module when multiple packages are passed in the ``name`` option (https://github.com/ansible-collections/community.general/pull/9181).