… tests/integration/targets/homebrew/tasks/formulae.yml