##### SUMMARY
<!--- Describe the change below, including rationale and design decisions -->

This is a rather big refactoring pull request but with rather big performance improvements.
**This makes this simple brew task 5-15x faster on my machine** by leveraging brew ability to work with multiple packages at a time.  Only 1 brew call is made to install N packages instead of N brew calls previously, greatly reducing the overall runtime.
```yml
- name: Install & upgrade brew packages
  community.general.homebrew:
    name: "{{ thibaut_brew_packages }}"
    state: latest
    update_homebrew: true
```

**Review commit per commit is highly recommended.** 
The first commit adds a few tests to increase coverage and validate more strictly the outputs. 
Then I tried to keep the commits contained and made sure tests are passing for each one of them.

##### ISSUE TYPE
- Refactoring Pull Request
- Test Pull Request

##### COMPONENT NAME
homebrew

##### ADDITIONAL INFORMATION
Here are the improvement I measured locally working with 15 packages (with package sizes from 2 to 250Mo)

- **It went from ~25s to ~1.5s** for the noop case (15 packages already up to date) 
- **It went from ~35s to ~7s** for the few updates case (5/15 packages updated) 

Any brew call takes ~1s because of the overhead of brew itself, this PR tries to cut most of it by doing batch brew calls instead of brew calls in a loop. The performance difference is even more noticeable for small packages were the actual install is fast compared to the brew overhead.

PS: If it makes sense, I'll be more than welcome to be marked as a maintainer of this module. I've seen you might be looking for one in https://github.com/ansible-collections/community.general/issues/8709#issuecomment-2335050203
PS2: This MR is a followup to #9022 were this optimization idea was first mentioned.