@UnknownPlatypus 2024-11-24
> You might want to take a look at ModuleHelper (not for this PR, of course):
Interesting, I'll look into it for a future PR. 
@russoz Any recommendation of existing module or PR using this helper i could take inspiration from ?